mapping function not working

+++ joystick.h
@@ -8,15 +8,15 @@ enum directions {
 };
 
 volatile struct joysticks {
-	volatile double x_val;
-	volatile double y_val;
+	volatile float x_val;
+	volatile float y_val;
 	volatile enum directions dir;
 };
 
 volatile struct joysticks fetchJoystick(void);
 volatile struct controllers fetchController(void);
 
-double joystickPercent(uint8_t val);
+float joystickPercent(uint8_t val);
 enum directions direction(uint8_t x_val, uint8_t y_val);
 
 void joystickTest(void);
